# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "B2C Marketing Automation API",
    "description": "Complete REST API for the B2C Marketing Automation Platform. Manage contacts, workflows, deliveries, segments, audiences, custom objects, and more.",
    "version": "2.0.0",
    "contact": {
      "name": "Platform Team"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "Current server"
    }
  ],
...
  "paths": {
...
    "/custom-objects": {
      "get": { "tags": ["Custom Objects"], "summary": "List custom objects", "responses": { "200": { "description": "All custom objects" } } },
      "post": {
        "tags": ["Custom Objects"],
        "summary": "Create custom object",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CustomObjectInput" } } } },
        "responses": { "201": { "description": "Object created" } }
      }
    },
...
    "/custom-objects/{id}": {
      "get": { "tags": ["Custom Objects"], "summary": "Get custom object", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "Object details" } } },
      "put": { "tags": ["Custom Objects"], "summary": "Update custom object", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "Updated" } } },
      "delete": { "tags": ["Custom Objects"], "summary": "Delete custom object", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "Deleted" } } }
    },
    "/custom-objects/{id}/ui": {
      "get": {
        "tags": ["Custom Objects"],
        "summary": "Get UI layout",
        "description": "Get draft or published UI builder layout.",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } },
          { "name": "mode", "in": "query", "schema": { "type": "string", "enum": ["draft", "published"], "default": "published" } }
        ],
        "responses": { "200": { "description": "Layout data" } }
      }
    },
    "/custom-objects/{id}/ui/draft": {
      "put": { "tags": ["Custom Objects"], "summary": "Save UI draft", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "Draft saved" } } }
    },
    "/custom-objects/{id}/ui/publish": {
      "post": { "tags": ["Custom Objects"], "summary": "Publish UI", "description": "Publish the current draft (previous published version is saved to version history).", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "Published" } } }
    },
    "/custom-objects/{id}/ui/versions": {
      "get": { "tags": ["Custom Objects"], "summary": "List UI versions", "description": "Get version history for a custom object's UI.", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "Version list" } } }
    },
    "/custom-objects/{id}/ui/versions/{versionId}": {
      "get": { "tags": ["Custom Objects"], "summary": "Get version layout", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }, { "name": "versionId", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Version layout data" } } }
    },
    "/custom-objects/{id}/ui/versions/{versionId}/restore": {
      "post": { "tags": ["Custom Objects"], "summary": "Restore version", "description": "Restore a previous version as the current draft.", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }, { "name": "versionId", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Version restored as draft" } } }
    },
    "/custom-objects/{id}/data": {
      "get": { "tags": ["Custom Objects"], "summary": "Get records", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "Record list" } } },
      "post": { "tags": ["Custom Objects"], "summary": "Add record", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "201": { "description": "Record created" } } }
    },
    "/custom-objects/{id}/import": {
      "post": { "tags": ["Custom Objects"], "summary": "Bulk import (CSV)", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "Import results" } } }
    },
    "/custom-objects/{id}/data/{recordId}": {
      "put": { "tags": ["Custom Objects"], "summary": "Update record", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }, { "name": "recordId", "in": "path", "required": true, "schema": { "type": "number" } }], "responses": { "200": { "description": "Record updated" } } },
      "delete": { "tags": ["Custom Objects"], "summary": "Delete record", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }, { "name": "recordId", "in": "path", "required": true, "schema": { "type": "number" } }], "responses": { "200": { "description": "Record deleted" } } }
    },
...
  },
  "components": {
    "schemas": {
...
      "CustomObjectInput": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "pattern": "^[a-z][a-z0-9_]*$" },
          "label": { "type": "string" },
          "description": { "type": "string" },
          "fields": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": { "type": "string" },
                "label": { "type": "string" },
                "type": { "type": "string", "enum": ["text", "number", "date", "datetime", "boolean"] },
                "is_required": { "type": "boolean" },
                "is_primary": { "type": "boolean" }
              },
              "required": ["name", "type"]
            }
          },
          "relationships": { "type": "array", "items": { "type": "object" } }
        },
        "required": ["name", "label", "fields"]
      }
    }
  }
}